:fire: Remove apps parts
