--- templates/Application/myApps.html.twig
@@ -1,101 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block body %}
-    <div class="row">
-        {% if starters is not empty %}
-        <div class="col-lg-9 col-xs-12">
-        {% else %}
-        <div class="col-lg-12 col-xs-12">
-        {% endif %}
-            <div class="row">
-                <div class="col-md-12">
-                    {% include '_partial/flash.html.twig' %}
-                    <h1>{{ 'main.menu.account.label'|trans }}</h1>
-                    {%  if passwordAges > 80 %}
-                    <div class="alert alert-danger" role="alert">
-                    {%  elseif passwordAges > 70 %}
-                    <div class="alert alert-warning" role="alert">
-                    {%  else %}
-                    <div class="alert alert-primary" role="alert">
-                    {% endif %}
-                        {{ 'my.account.label.lastset'|trans }}:
-                        <time class="timeago" datetime="{{ user.passwordLastSetTimestamp|date("Y-m-d H:i:s") }}">{{ user.passwordLastSetTimestamp|date("d/m/Y") }}</time>
-                        <br/><small>{{ 'my.account.label.password.recommendation'|trans }}</small>
-                    </div>
-                </div>
-                <div class="col-lg-6">
-                    <div class="card">
-                        <h4 class="card-header">
-                            <i class="flag-icon flag-icon-{{ country|lower }}" title="{{ user.department }}" alt="{{ user.department }}"></i> {{ user.firstname }} {{ user.lastname }}
-                        </h4>
-                        <div class="card-body">
-                            <h5 class="card-title">{{ user.description }}</h5>
-                            <p class="card-text">
-                                <i class="fa fa-envelope" aria-hidden="true"></i> {{ user.userprincipalname }}
-                                {% if user.mobileNumber is not empty %}
-                                    <br>
-                                    <i class="fa fa-mobile" aria-hidden="true"></i> {{ user.mobileNumber }}
-                                {% elseif user.telephoneNumber is not empty %}
-                                    <br>
-                                    <i class="fa fa-phone" aria-hidden="true"></i> {{ user.telephoneNumber }}
-                                {% endif %}
-                            </p>
-                            <div class="btn-group" role="group" aria-label="Actions">
-                                <a href="{{ url('account_change_password') }}" class="btn btn-warning"><i class="fa fa-key"></i> {{ 'btn.change.password.label'|trans }}</a>
-                                <a href="{{ url('contact_own_country') }}" class="btn btn-info"><i class="fa fa-phone"></i> {{ 'btn.contact.own.country.label'|trans }} <i class="flag-icon flag-icon-{{ app.user.username|userCountry|lower }}"></i></a>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-md-12">
-                    <h1>{{ 'main.menu.application.label'|trans }}</h1>
-                    <div class="card-columns">
-                        {% for application in applications %}
-                            <div class="card text-center">
-                                <div class="card-body">
-                                    <h5 class="card-title">{{ application.name }}</h5>
-                                    {%  if application.description is not empty %}
-                                        <p class="card-text text-left small">{{ application.description|raw }}</p>
-                                    {% endif %}
-                                    <a href="{{ application.link }}" target="_blank" class="btn btn-primary"><i class="fa fa-eye"></i> {{ 'application.btn.link.label'|trans|upper }}</a>
-                                </div>
-                            </div>
-                        {% endfor %}
-                    </div>
-                </div>
-            </div>
-        </div>
-        {% if starters is not empty %}
-        <div class="col-lg-3 col-xs-12">
-            <div class="row">
-                <ul class="list-group col-lg-12">
-                    <li class="list-group-item list-group-item-info"><i class="fa fa-users" aria-hidden="true"></i> {{ 'application.new_collegues.title.h3'|trans }}</li>
-                    {% for collegue in starters %}
-                        <li class="list-group-item d-flex justify-content-between align-items-start">
-                        <span>
-                            {{ collegue.countryFlag|raw }} {{ collegue.displayName }}
-                            <br>
-                            <small>
-                                {{ 'application.new_collegues.start_date'|trans }}: {{ collegue.DateContractStart|date('d/m/Y') }}<br>
-                                <i class="fa fa-briefcase" aria-hidden="true"></i> {{ collegue.function }}
-                            </small>
-                        </span>
-                            <span>
-                            {% if is_granted('ROLE_ADMIN') %}
-                                <a href="{{ path('account_detail', {'id': collegue.employeeId }) }}" class="btn btn-outline-success btn-sm"><i class="fa fa-eye"></i> {{ 'Show detail'|trans }}</a>
-                            {% endif %}
-                        </span>
-                        </li>
-                    {% endfor %}
-                </ul>
-            </div>
-        </div>
-        {% endif %}
-    </div>
-{% endblock %}
-
-{% block title %}
-    {{ 'main.menu.application.label'|trans }} - {{ parent() }}
-{% endblock %}
