+++ templates/_partial/menu.html.twig
@@ -1,5 +1,5 @@
-<li class="nav-item {% if (app.request.get('_route') == 'homepage') or (app.request.get('_route') == 'application-my-apps') %}active{% endif %}">
-    <a href="{{ path('application-my-apps') }}" class="nav-link">{{ 'main.menu.application.label'|trans }}</a>
+<li class="nav-item {% if (app.request.get('_route') == 'homepage') %}active{% endif %}">
+    <a href="{{ path('homepage') }}" class="nav-link"><i class="fa fa-home"></i> {{ 'main.menu.homepage.label'|trans }}</a>
 </li>
 
 {% if is_granted('ROLE_ADMIN') %}
